+++ firestore-bigquery-export/scripts/import/src/index.ts
@@ -19,7 +19,6 @@
 import * as firebase from "firebase-admin";
 import * as program from "commander";
 import * as fs from "fs";
-import * as inquirer from "inquirer";
 import * as util from "util";
 import * as filenamify from "filenamify";
 import { runMultiThread } from "./run";
@@ -30,6 +29,7 @@ import {
   FirestoreBigQueryEventHistoryTracker,
   FirestoreDocumentChangeEvent,
 } from "@firebaseextensions/firestore-bigquery-change-tracker";
+import { parseConfig } from "./config";
 
 // For reading cursor position.
 const exists = util.promisify(fs.exists);
@@ -37,70 +37,7 @@ const write = util.promisify(fs.writeFile);
 const read = util.promisify(fs.readFile);
 const unlink = util.promisify(fs.unlink);
 
-const BIGQUERY_VALID_CHARACTERS = /^[a-zA-Z0-9_]+$/;
-const FIRESTORE_VALID_CHARACTERS = /^[^({|})]+$/;
-
-const PROJECT_ID_MAX_CHARS = 6144;
-const FIRESTORE_COLLECTION_NAME_MAX_CHARS = 6144;
-const BIGQUERY_RESOURCE_NAME_MAX_CHARS = 1024;
-
 const FIRESTORE_DEFAULT_DATABASE = "(default)";
-
-const validateInput = (
-  value: string,
-  name: string,
-  regex: RegExp,
-  sizeLimit: number
-) => {
-  if (!value || value === "" || value.trim() === "") {
-    return `Please supply a ${name}`;
-  }
-  if (value.length >= sizeLimit) {
-    return `${name} must be at most ${sizeLimit} characters long`;
-  }
-  if (!value.match(regex)) {
-    return `The ${name} does not match the regular expression provided`;
-  }
-  return true;
-};
-
-const validateBatchSize = (value: string) => {
-  return parseInt(value, 10) > 0;
-};
-
-const validateLocation = (value: string) => {
-  const index = [
-    "asia-east1",
-    "asia-east2",
-    "asia-northeast1",
-    "asia-northeast2",
-    "asia-northeast3",
-    "asia-south1",
-    "asia-southeast1",
-    "asia-southeast2",
-    "australia-southeast1",
-    "eu",
-    "europe-central2",
-    "europe-north1",
-    "europe-west1",
-    "europe-west3",
-    "europe-west2",
-    "europe-west4",
-    "europe-west6",
-    "northamerica-northeast1",
-    "southamerica-east1",
-    "us",
-    "us-central1",
-    "us-east1",
-    "us-east4",
-    "us-west1",
-    "us-west2",
-    "us-west3",
-    "us-west4",
-  ].indexOf(value.toLowerCase());
-
-  return index !== -1;
-};
 
 const packageJson = require("../package.json");
 
@@ -146,138 +83,56 @@ program
   .option(
     "-m, --multi-threaded [true|false]",
     "Whether to run standard or multi-thread import version"
+  )
+  .option(
+    "-u, --use-new-snapshot-query-syntax [true|false]",
+    "Whether to use updated latest snapshot query"
+  )
+  .option(
+    "-e, --use-emulator [true|false]",
+    "Whether to use updated latest snapshot query"
   );
 
-const questions = [
-  {
-    message: "What is your Firebase project ID?",
-    name: "project",
-    type: "input",
-    validate: (value) =>
-      validateInput(
-        value,
-        "project ID",
-        FIRESTORE_VALID_CHARACTERS,
-        PROJECT_ID_MAX_CHARS
-      ),
-  },
-  {
-    message:
-      "What is the path of the the Cloud Firestore Collection you would like to import from? " +
-      "(This may, or may not, be the same Collection for which you plan to mirror changes.)",
-    name: "sourceCollectionPath",
-    type: "input",
-    validate: (value) =>
-      validateInput(
-        value,
-        "collection path",
-        FIRESTORE_VALID_CHARACTERS,
-        FIRESTORE_COLLECTION_NAME_MAX_CHARS
-      ),
-  },
-  {
-    message: "Would you like to import documents via a Collection Group query?",
-    name: "queryCollectionGroup",
-    type: "confirm",
-    default: false,
-  },
-  {
-    message:
-      "What is the ID of the BigQuery dataset that you would like to use? (A dataset will be created if it doesn't already exist)",
-    name: "dataset",
-    type: "input",
-    validate: (value) =>
-      validateInput(
-        value,
-        "dataset",
-        BIGQUERY_VALID_CHARACTERS,
-        BIGQUERY_RESOURCE_NAME_MAX_CHARS
-      ),
-  },
-  {
-    message:
-      "What is the identifying prefix of the BigQuery table that you would like to import to? (A table will be created if one doesn't already exist)",
-    name: "table",
-    type: "input",
-    validate: (value) =>
-      validateInput(
-        value,
-        "table",
-        BIGQUERY_VALID_CHARACTERS,
-        BIGQUERY_RESOURCE_NAME_MAX_CHARS
-      ),
-  },
-  {
-    message:
-      "How many documents should the import stream into BigQuery at once?",
-    name: "batchSize",
-    type: "input",
-    default: 300,
-    validate: validateBatchSize,
-  },
-  {
-    message: "Where would you like the BigQuery dataset to be located?",
-    name: "datasetLocation",
-    type: "input",
-    default: "us",
-    validate: validateLocation,
-  },
-  {
-    message: "Would you like to run the import across multiple threads?",
-    name: "multiThreaded",
-    type: "confirm",
-    default: false,
-  },
-];
-
-type CliConfig = {
-  kind: "CONFIG";
-  projectId: string;
-  sourceCollectionPath: string;
-  datasetId: string;
-  tableId: string;
-  batchSize: string;
-  queryCollectionGroup: boolean;
-  datasetLocation: string;
-  multiThreaded: boolean;
-};
-
-type CliConfigError = {
-  kind: "ERROR";
-  errors: string[];
-};
-
 const run = async (): Promise<number> => {
-  const parsed = await parseConfig();
-  if (parsed.kind === "ERROR") {
-    parsed.errors.forEach((e) => console.error(`[ERROR] ${e}`));
+  const config = await parseConfig();
+  if (config.kind === "ERROR") {
+    config.errors?.forEach((e) => console.error(`[ERROR] ${e}`));
     process.exit(1);
   }
 
   const {
     projectId,
     sourceCollectionPath,
-    queryCollectionGroup,
     datasetId,
     tableId,
     batchSize,
+    queryCollectionGroup,
     datasetLocation,
     multiThreaded,
-  }: CliConfig = parsed;
+    useNewSnapshotQuerySyntax,
+    useEmulator,
+  } = config;
 
-  if (multiThreaded) return runMultiThread();
+  if (useEmulator) {
+    console.log("Using emulator");
+    process.env.FIRESTORE_EMULATOR_HOST = "localhost:8080";
+  }
 
-  const batch = parseInt(batchSize);
-  const rawChangeLogName = `${tableId}_raw_changelog`;
   // Initialize Firebase
+  // This uses applicationDefault to authenticate
+  // Please see https://cloud.google.com/docs/authentication/production
   firebase.initializeApp({
     credential: firebase.credential.applicationDefault(),
     databaseURL: `https://${projectId}.firebaseio.com`,
   });
-  // Set project ID so it can be used in BigQuery intialization
+
+  // Set project ID, so it can be used in BigQuery initialization
   process.env.PROJECT_ID = projectId;
   process.env.GOOGLE_CLOUD_PROJECT = projectId;
 
+  const rawChangeLogName = `${tableId}_raw_changelog`;
+
+  if (multiThreaded) return runMultiThread(config, rawChangeLogName);
   // We pass in the application-level "tableId" here. The tracker determines
   // the name of the raw changelog from this field.
   const dataSink = new FirestoreBigQueryEventHistoryTracker({
@@ -285,6 +140,7 @@ const run = async (): Promise<number> => {
     datasetId: datasetId,
     datasetLocation,
     wildcardIds: queryCollectionGroup,
+    useNewSnapshotQuerySyntax,
   });
 
   console.log(
@@ -327,7 +183,7 @@ const run = async (): Promise<number> => {
       query = firebase.firestore().collection(sourceCollectionPath);
     }
 
-    query = query.limit(batch);
+    query = query.limit(batchSize);
 
     if (cursor) {
       query = query.startAfter(cursor);
@@ -356,6 +212,7 @@ const run = async (): Promise<number> => {
   try {
     await unlink(cursorPositionFile);
   } catch (e) {
+    console.log(e);
     console.log(
       `Error unlinking journal file ${cursorPositionFile} after successful import: ${e.toString()}`
     );
@@ -363,77 +220,6 @@ const run = async (): Promise<number> => {
 
   return totalRowsImported;
 };
-
-async function parseConfig(): Promise<CliConfig | CliConfigError> {
-  program.parse(process.argv);
-
-  if (program.nonInteractive) {
-    const errors = [];
-    if (program.project === undefined) {
-      errors.push("Project is not specified.");
-    }
-    if (program.sourceCollectionPath === undefined) {
-      errors.push("SourceCollectionPath is not specified.");
-    }
-    if (program.dataset === undefined) {
-      errors.push("Dataset ID is not specified.");
-    }
-    if (program.tableNamePrefix === undefined) {
-      errors.push("TableNamePrefix is not specified.");
-    }
-    if (program.queryCollectionGroup === undefined) {
-      errors.push("QueryCollectionGroup is not specified.");
-    }
-    if (program.batchSize === undefined) {
-      errors.push("BatchSize is not specified.");
-    }
-    if (program.datasetLocation === undefined) {
-      errors.push("DatasetLocation is not specified.");
-    }
-    if (!validateBatchSize(program.batchSize)) {
-      errors.push("Invalid batch size.");
-    }
-
-    if (errors.length !== 0) {
-      program.outputHelp();
-      return { kind: "ERROR", errors };
-    }
-
-    return {
-      kind: "CONFIG",
-      projectId: program.project,
-      sourceCollectionPath: program.sourceCollectionPath,
-      datasetId: program.dataset,
-      tableId: program.tableNamePrefix,
-      batchSize: program.batchSize,
-      queryCollectionGroup: program.queryCollectionGroup === "true",
-      datasetLocation: program.datasetLocation,
-      multiThreaded: program.multiThreaded === "true",
-    };
-  }
-  const {
-    project,
-    sourceCollectionPath,
-    dataset,
-    table,
-    batchSize,
-    queryCollectionGroup,
-    datasetLocation,
-    multiThreaded,
-  } = await inquirer.prompt(questions);
-
-  return {
-    kind: "CONFIG",
-    projectId: project,
-    sourceCollectionPath: sourceCollectionPath,
-    datasetId: dataset,
-    tableId: table,
-    batchSize: batchSize,
-    queryCollectionGroup: queryCollectionGroup,
-    datasetLocation: datasetLocation,
-    multiThreaded: multiThreaded,
-  };
-}
 
 run()
   .then((rowCount) => {